requirements.txt
# Python dependencies for Vexo Music Bot

# Discord
discord.py[voice]>=2.3.0

# YouTube & Spotify
yt-dlp>=2024.1.0
ytmusicapi>=1.3.0
spotipy>=2.23.0

# Metadata (Fallbacks)
discogs-client>=2.3.0
musicbrainzngs>=0.7.1

# Database & Async
aiosqlite>=0.19.0
aiohttp>=3.9.0
python-dotenv>=1.0.0

# System & Audio Processing
PyNaCl>=1.5.0
psutil>=5.9.0
